Rename 'gltfMetadata' to 'gltfExtensions'
