+++ src/gltfExtensions/EXTStructuralMetadata.ts
