+++ src/gltfExtensions/InstanceFeatures.ts
