+++ src/gltfExtensions/InstanceFeaturesUtils.ts
