+++ src/gltfExtensions/MeshFeatures.ts
