+++ src/gltfExtensions/MeshFeaturesUtils.ts
