+++ src/gltfExtensions/StringBuilder.ts
