+++ src/gltfExtensions/StructuralMetadata.ts
