+++ src/gltfExtensions/StructuralMetadataPropertyTables.ts
