+++ src/gltfExtensions/StructuralMetadataUtils.ts
